+++ include/Socket/SSocket.hpp
@@ -6,7 +6,7 @@
 /*   By: agimi <agimi@student.1337.ma>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 10:02:13 by agimi             #+#    #+#             */
-/*   Updated: 2024/02/14 12:34:16 by agimi            ###   ########.fr       */
+/*   Updated: 2024/02/14 14:52:06 by agimi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ namespace wbs
 		SSocket &operator=(const SSocket &);
 
 	public:
-		SSocket(int, int, int, int, u_long);
+		SSocket(int, int, int, int, unsigned int);
 		~SSocket();
 
 		virtual int connect() = 0;
